Move `State` object from handlers to internals
